requirements.txt
aiogram==3.*

selenium>=4.23
beautifulsoup4>=4.12.3
lxml>=5.3.0
html5lib>=1.1

python-dotenv>=1.0.1
requests>=2.32.3
urllib3>=2.2.3

aiofiles>=23.2.1
